模拟其他第三方接口进行重构

+++ HousePriceApi/src/main/java/com/yunfang/utils/BigDecimalUtils.java
@@ -14,10 +14,10 @@ public class BigDecimalUtils {
 	 *            加数
 	 * @return 两个参数的和
 	 */
-	public static double add(double v1, double v2) {
+	public static BigDecimal add(double v1, double v2) {
 		BigDecimal b1 = new BigDecimal(Double.toString(v1));
 		BigDecimal b2 = new BigDecimal(Double.toString(v2));
-		return b1.add(b2).doubleValue();
+		return b1.add(b2);
 	}
 
 	/**
@@ -29,10 +29,10 @@ public class BigDecimalUtils {
 	 *            减数
 	 * @return 两个参数的差
 	 */
-	public static double sub(double v1, double v2) {
+	public static BigDecimal sub(double v1, double v2) {
 		BigDecimal b1 = new BigDecimal(Double.toString(v1));
 		BigDecimal b2 = new BigDecimal(Double.toString(v2));
-		return b1.subtract(b2).doubleValue();
+		return b1.subtract(b2);
 	}
 
 	/**
@@ -44,10 +44,10 @@ public class BigDecimalUtils {
 	 *            乘数
 	 * @return 两个参数的积
 	 */
-	public static double mul(double v1, double v2) {
+	public static BigDecimal mul(double v1, double v2) {
 		BigDecimal b1 = new BigDecimal(Double.toString(v1));
 		BigDecimal b2 = new BigDecimal(Double.toString(v2));
-		return b1.multiply(b2).doubleValue();
+		return b1.multiply(b2);
 	}
 
 	/**
@@ -59,7 +59,7 @@ public class BigDecimalUtils {
 	 *            除数
 	 * @return 两个参数的商
 	 */
-	public static double div(double v1, double v2) {
+	public static BigDecimal div(double v1, double v2) {
 		return div(v1, v2, DEF_DIV_SCALE);
 	}
 
@@ -74,14 +74,14 @@ public class BigDecimalUtils {
 	 *            表示表示需要精确到小数点以后几位。
 	 * @return 两个参数的商
 	 */
-	public static double div(double v1, double v2, int scale) {
+	public static BigDecimal div(double v1, double v2, int scale) {
 		if (scale < 0) {
 			throw new IllegalArgumentException(
 					"The scale must be a positive integer or zero");
 		}
 		BigDecimal b1 = new BigDecimal(Double.toString(v1));
 		BigDecimal b2 = new BigDecimal(Double.toString(v2));
-		return b1.divide(b2, scale, BigDecimal.ROUND_HALF_UP).doubleValue();
+		return b1.divide(b2, scale, BigDecimal.ROUND_HALF_UP);
 	}
 
 	/**
@@ -93,14 +93,14 @@ public class BigDecimalUtils {
 	 *            小数点后保留几位
 	 * @return 四舍五入后的结果
 	 */
-	public static double round(double v, int scale) {
+	public static BigDecimal round(double v, int scale) {
 		if (scale < 0) {
 			throw new IllegalArgumentException(
 					"The scale must be a positive integer or zero");
 		}
 		BigDecimal b = new BigDecimal(Double.toString(v));
 		BigDecimal one = new BigDecimal("1");
-		return b.divide(one, scale, BigDecimal.ROUND_HALF_UP).doubleValue();
+		return b.divide(one, scale, BigDecimal.ROUND_HALF_UP);
 	}
 	
 	/**
